# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FindingResponse": {
        "properties": {
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line End"
          },
          "line_start": {
            "title": "Line Start",
            "type": "integer"
          },
          "matched_content_sha256": {
            "title": "Matched Content Sha256",
            "type": "string"
          },
          "penalty": {
            "title": "Penalty",
            "type": "integer"
          },
          "remediation_link": {
            "title": "Remediation Link",
            "type": "string"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status_at_scan": {
            "enum": [
              "shadow",
              "active"
            ],
            "title": "Status At Scan",
            "type": "string"
          },
          "sub_score": {
            "enum": [
              "security",
              "supply_chain",
              "maintenance",
              "transparency",
              "community"
            ],
            "title": "Sub Score",
            "type": "string"
          }
        },
        "required": [
          "id",
          "rule_id",
          "severity",
          "sub_score",
          "penalty",
          "status_at_scan",
          "file_path",
          "line_start",
          "matched_content_sha256",
          "remediation_link",
          "rubric_version"
        ],
        "title": "FindingResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ScanReportSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "data"
        ],
        "title": "ListEnvelope",
        "type": "object"
      },
      "ScanReportDetail": {
        "properties": {
          "aggregate_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Aggregate Score",
            "type": "integer"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingResponse"
            },
            "title": "Findings",
            "type": "array"
          },
          "github_url": {
            "title": "Github Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "ref_sha": {
            "title": "Ref Sha",
            "type": "string"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          },
          "scanned_at": {
            "format": "date-time",
            "title": "Scanned At",
            "type": "string"
          },
          "score_breakdown": {
            "additionalProperties": true,
            "title": "Score Breakdown",
            "type": "object"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "default": "completed",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "sub_scores": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Sub Scores",
            "type": "object"
          },
          "tier": {
            "enum": [
              "green",
              "yellow",
              "orange",
              "red",
              "unscoped"
            ],
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "id",
          "github_url",
          "slug",
          "display_name",
          "aggregate_score",
          "tier",
          "sub_scores",
          "score_breakdown",
          "findings",
          "scanned_at",
          "rubric_version",
          "engine_version",
          "latency_ms",
          "source",
          "ref_sha"
        ],
        "title": "ScanReportDetail",
        "type": "object"
      },
      "ScanReportSummary": {
        "properties": {
          "aggregate_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Aggregate Score",
            "type": "integer"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "findings_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Findings Count",
            "type": "integer"
          },
          "github_url": {
            "title": "Github Url",
            "type": "string"
          },
          "id": {
            "description": "Scan UUID rendered as string.",
            "title": "Id",
            "type": "string"
          },
          "scanned_at": {
            "format": "date-time",
            "title": "Scanned At",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tier": {
            "enum": [
              "green",
              "yellow",
              "orange",
              "red",
              "unscoped"
            ],
            "title": "Tier",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "github_url",
          "slug",
          "aggregate_score",
          "tier",
          "scanned_at"
        ],
        "title": "ScanReportSummary",
        "type": "object"
      },
      "ScanSubmitRequest": {
        "properties": {
          "github_url": {
            "description": "Public github.com repository URL.",
            "title": "Github Url",
            "type": "string"
          },
          "rescan": {
            "default": false,
            "description": "If true, bypass idempotency cache and queue a new scan.",
            "title": "Rescan",
            "type": "boolean"
          }
        },
        "required": [
          "github_url"
        ],
        "title": "ScanSubmitRequest",
        "type": "object"
      },
      "ScanSubmitResponse": {
        "properties": {
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "rubric_version",
          "submitted_at"
        ],
        "title": "ScanSubmitResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Every AI skill, independently scanned \u2014 public, free, Apache-2.0 trust-scoring for skills, MCP servers, hooks, and plugins.",
    "title": "SaferSkills API",
    "version": "0.0.0+foundation"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scans": {
      "get": {
        "description": "Slim summary list \u2014 JOINs scans + catalog_items, projects ScanReportSummary.",
        "operationId": "list_scans_api_v1_scans_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "submission",
                    "ingestion",
                    "rescan_drift",
                    "rescan_appeal"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "green",
                    "yellow",
                    "orange",
                    "red",
                    "unscoped"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "created_at_desc",
              "enum": [
                "created_at_desc",
                "created_at_asc",
                "score_desc",
                "score_asc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List recent scans.",
        "tags": [
          "scans"
        ]
      },
      "post": {
        "operationId": "submit_scan_api_v1_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a public GitHub URL for scanning.",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}": {
      "get": {
        "operationId": "get_scan_api_v1_scans__scan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the full scan report for a given scan ID.",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/events": {
      "get": {
        "description": "Server-Sent Events stream of scan progress.\n\nWire format: each event is `{event_seq, stage, completion_pct, status,\npayload, timestamp}` JSON. Clients pass `Last-Event-ID: <scan_id>-<seq>`\nto resume \u2014 we replay rows from `scan_events` whose `event_seq > seq`,\nthen subscribe to LISTEN for live deltas.",
        "operationId": "scan_events_api_v1_scans__scan_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "SSE progress stream for a scan.",
        "tags": [
          "scans"
        ]
      }
    }
  }
}